refactor: improve type safety and API ergonomics
This commit refactors the `:module:library:preferences` module to enhance type safety and improve its public API. The `Preferences` interface is now more explicit about types by using `KClass` during data retrieval.

- **Type-Safe API:**
    - The `Preferences` interface's `getOrThrow` and `getOrDefault` methods are replaced by a single, type-safe `get<T : Any>(key: String, kClass: KClass<T>): T?` method.
    - A new `UnsupportedTypeException` is thrown when a value is retrieved with a different type than it was stored with, preventing runtime casting errors.
    - `SerializedType` enum now holds a `KClass` reference for each type, enabling compile-time type validation.

- **API Ergonomics & Extension Functions:**
    - A new `PreferencesExt.kt` file is added, containing inline extension functions (`getOrDefault`, `getOrThrow`, and `edit`) that use reified types. This provides a more convenient and idiomatic Kotlin API for consumers.
    - The `edit` extension simplifies the common get-modify-set pattern.

- **Sample Feature Update:**
    - The `StorageDemoViewModel` is updated to use the new type-safe extension functions, demonstrating their usage and cleaning up the logic for managing counters and user-interactive values.

- **Testing:**
    - Unit tests for `DbPreferences` (renamed to `PreferencesTest`) have been updated to reflect the new API and verify the type-checking logic.
    - Added tests for `SerializedType` to confirm the correctness of its `KClass` associations.
    - The `PrimitivesSerializerTest` is updated to check for `IllegalArgumentException` on strict boolean decoding.

+++ module/feature/sample/src/commonMain/kotlin/kmp/template/feature/sample/presentation/storage/StorageDemoViewState.kt
@@ -5,8 +5,8 @@ import kmp.template.design.component.screenstate.ScreenStateUiModel.Loading
 
 internal data class StorageDemoViewState(
     val screenState: ScreenStateUiModel = Loading(),
-    val viewModelInitCounter: Int = 0,
-    val composeLaunchCounter: Int = 0,
-    val userInteractValue: Long = 0L,
+    val viewModelInitCounter: Long = 0L,
+    val composeLaunchCounter: Long = 0L,
+    val userInteractValue: Int = 0,
     val userInteractKeyExist: Boolean = false
 )
